refactor(mcp): state.rs decomposition + symbols.rs cleanup — net -594 lines
Major structural cleanup addressing ADR-0001 and ADR-007:

1. State.rs God Object decomposition (ADR-007)
   - Extract session_runtime.rs (session-scoped state management)
   - Extract watcher_health.rs (file watcher health tracking)
   - state.rs: 127 lines removed, concerns distributed to sub-modules

2. tools/symbols.rs cleanup
   - 546 lines removed — dead code paths, redundant query expansion,
     over-engineered semantic reranking that was dominated by the
     engine-level scoring pipeline
   - Remaining code is focused dispatch + response assembly

3. LSP registry consolidation (ADR-0001 single-source)
   - lsp/registry.rs: deduplicated default command/args registry
   - lsp/tests.rs: registry tests consolidated
   - lib.rs: cleaner re-exports

4. Dispatch + impact_reports adjustments
   - dispatch.rs: aligned with new state sub-module imports
   - impact_reports.rs: simplified query path

Net change: +115 / -709 = **-594 lines** (code deletion > addition).
This is the largest single code-quality improvement in the v1.7 arc.

262 engine + 172 MCP tests pass.
