+++ crates/codelens-engine/src/lsp/mod.rs
@@ -4,7 +4,10 @@ pub mod registry;
 pub(crate) mod session;
 pub mod types;
 
-pub use registry::{LSP_RECIPES, LspRecipe, LspStatus, check_lsp_status, get_lsp_recipe};
+pub use registry::{
+    LSP_RECIPES, LspRecipe, LspStatus, check_lsp_status, default_lsp_args_for_command,
+    default_lsp_command_for_extension, default_lsp_command_for_path, get_lsp_recipe,
+};
 pub use session::LspSessionPool;
 pub use types::{
     LspDiagnostic, LspDiagnosticRequest, LspReference, LspRenamePlan, LspRenamePlanRequest,
