+++ crates/codelens-mcp/src/state.rs
@@ -2,7 +2,6 @@
 use codelens_engine::EmbeddingEngine;
 use codelens_engine::{FileWatcher, GraphCache, LspSessionPool, ProjectRoot, SymbolIndex};
 use std::collections::HashMap;
-use std::fs;
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex, OnceLock};
 
@@ -18,7 +17,10 @@ use crate::runtime_types::JobLifecycle;
 use crate::telemetry::ToolMetricsRegistry;
 use crate::tool_defs::{ToolPreset, ToolProfile, ToolSurface};
 use serde_json::Value;
-use std::collections::VecDeque;
+
+mod project_runtime;
+mod session_runtime;
+mod watcher_health;
 
 const WATCHER_RECENT_FAILURE_WINDOW_SECS: i64 = 15 * 60;
 /// Default preflight TTL: 10 minutes. Override via CODELENS_PREFLIGHT_TTL_SECS.
@@ -62,70 +64,7 @@ fn normalize_path_for_project(project_root: &Path, path: &str) -> String {
 
 const PROJECT_CONTEXT_CACHE_LIMIT: usize = 4;
 
-/// Holds project-specific resources that can be reused across rebinds.
-struct ProjectRuntimeContext {
-    project: ProjectRoot,
-    symbol_index: Arc<SymbolIndex>,
-    graph_cache: Arc<GraphCache>,
-    lsp_pool: Arc<LspSessionPool>,
-    memories_dir: PathBuf,
-    analysis_dir: PathBuf,
-    audit_dir: PathBuf,
-    /// Keeps the watcher alive so it continues to receive file-system events.
-    watcher: Option<FileWatcher>,
-}
-
-#[derive(Default)]
-struct ProjectContextCache {
-    entries: HashMap<String, Arc<ProjectRuntimeContext>>,
-    access_order: VecDeque<String>,
-}
-
-impl ProjectContextCache {
-    fn get(&mut self, scope: &str) -> Option<Arc<ProjectRuntimeContext>> {
-        let context = self.entries.get(scope).cloned()?;
-        self.touch(scope);
-        Some(context)
-    }
-
-    fn insert(&mut self, scope: String, context: Arc<ProjectRuntimeContext>) {
-        self.entries.insert(scope.clone(), context);
-        self.touch(&scope);
-    }
-
-    fn touch(&mut self, scope: &str) {
-        self.access_order.retain(|entry| entry != scope);
-        self.access_order.push_back(scope.to_owned());
-    }
-
-    fn evict_until_within_limit(
-        &mut self,
-        limit: usize,
-        protected_scopes: &[&str],
-    ) -> Vec<Arc<ProjectRuntimeContext>> {
-        let mut evicted = Vec::new();
-        while self.entries.len() > limit {
-            let Some(oldest) = self.access_order.pop_front() else {
-                break;
-            };
-            if protected_scopes.iter().any(|scope| *scope == oldest) {
-                self.access_order.push_back(oldest);
-                if self.access_order.iter().all(|scope| {
-                    protected_scopes
-                        .iter()
-                        .any(|protected| protected == &scope.as_str())
-                }) {
-                    break;
-                }
-                continue;
-            }
-            if let Some(context) = self.entries.remove(&oldest) {
-                evicted.push(context);
-            }
-        }
-        evicted
-    }
-}
+use self::project_runtime::{ProjectContextCache, ProjectRuntimeContext};
 
 pub(crate) struct AppState {
     // Default project (set at startup, immutable)
@@ -256,111 +195,32 @@ impl AppState {
     }
 
     fn active_project_context(&self) -> Option<Arc<ProjectRuntimeContext>> {
-        self.project_override
-            .read()
-            .unwrap_or_else(|p| p.into_inner())
-            .as_ref()
-            .cloned()
+        project_runtime::active_project_context(self)
     }
 
     fn build_project_runtime_context(
         project: ProjectRoot,
         start_watcher: bool,
     ) -> anyhow::Result<ProjectRuntimeContext> {
-        let symbol_index = Arc::new(SymbolIndex::new(project.clone()));
-        if symbol_index
-            .stats()
-            .map(|s| s.indexed_files == 0)
-            .unwrap_or(true)
-        {
-            let _ = symbol_index.refresh_all();
-        }
-        let graph_cache = Arc::new(GraphCache::new(30));
-        let memories_dir = project.as_path().join(".codelens").join("memories");
-        let analysis_dir = project.as_path().join(".codelens").join("analysis-cache");
-        let audit_dir = project.as_path().join(".codelens").join("audit");
-        let _ = fs::create_dir_all(&memories_dir);
-        let _ = fs::create_dir_all(&analysis_dir);
-        let _ = fs::create_dir_all(analysis_dir.join("jobs"));
-        let _ = fs::create_dir_all(&audit_dir);
-        let lsp_pool = Arc::new(LspSessionPool::new(project.clone()));
-        let watcher = if start_watcher {
-            FileWatcher::start(
-                project.as_path(),
-                Arc::clone(&symbol_index),
-                Arc::clone(&graph_cache),
-            )
-            .ok()
-        } else {
-            None
-        };
-        Ok(ProjectRuntimeContext {
-            project,
-            symbol_index,
-            graph_cache,
-            lsp_pool,
-            memories_dir,
-            analysis_dir,
-            audit_dir,
-            watcher,
-        })
+        project_runtime::build_project_runtime_context(project, start_watcher)
     }
 
     fn activate_project_context(&self, context: Option<Arc<ProjectRuntimeContext>>) {
-        *self
-            .project_override
-            .write()
-            .unwrap_or_else(|p| p.into_inner()) = context.clone();
-        let analysis_dir = context
-            .as_ref()
-            .map(|override_ctx| override_ctx.analysis_dir.clone())
-            .unwrap_or_else(|| self.default_analysis_dir.clone());
-        self.artifact_store.set_analysis_dir(analysis_dir.clone());
-        self.job_store.set_jobs_dir(analysis_dir.join("jobs"));
-        self.artifact_store.clear();
-        self.job_store.clear();
-        self.clear_recent_preflights();
-        #[cfg(feature = "semantic")]
-        self.reset_embedding();
-        self.artifact_store.cleanup_stale_dirs(Self::now_ms());
-        let scope = self.current_project_scope();
-        self.job_store
-            .cleanup_stale_files(Self::now_ms(), Some(&scope));
-    }
-
-    #[cfg(feature = "http")]
-    fn http_session_state(
-        &self,
-        session: &crate::session_context::SessionRequestContext,
-    ) -> Option<Arc<crate::server::session::SessionState>> {
-        if session.is_local() {
-            return None;
-        }
-        self.session_store
-            .as_ref()
-            .and_then(|store| store.get(&session.session_id))
+        project_runtime::activate_project_context(self, context)
     }
 
     pub(crate) fn execution_surface(
         &self,
         _session: &crate::session_context::SessionRequestContext,
     ) -> ToolSurface {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            return session_state.surface();
-        }
-        *self.surface()
+        session_runtime::execution_surface(self, _session)
     }
 
     pub(crate) fn execution_token_budget(
         &self,
         _session: &crate::session_context::SessionRequestContext,
     ) -> usize {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            return session_state.token_budget();
-        }
-        self.token_budget()
+        session_runtime::execution_token_budget(self, _session)
     }
 
     #[cfg(feature = "http")]
@@ -370,12 +230,7 @@ impl AppState {
         surface: ToolSurface,
         budget: usize,
     ) {
-        if let Some(store) = &self.session_store
-            && let Some(session) = store.get(session_id)
-        {
-            session.set_surface(surface);
-            session.set_token_budget(budget);
-        }
+        session_runtime::set_session_surface_and_budget(self, session_id, surface, budget)
     }
 
     pub(crate) fn push_recent_tool_for_session(
@@ -383,23 +238,14 @@ impl AppState {
         _session: &crate::session_context::SessionRequestContext,
         name: &str,
     ) {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            session_state.push_recent_tool(name);
-            return;
-        }
-        self.push_recent_tool(name);
+        session_runtime::push_recent_tool_for_session(self, _session, name);
     }
 
     pub(crate) fn recent_tools_for_session(
         &self,
         _session: &crate::session_context::SessionRequestContext,
     ) -> Vec<String> {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            return session_state.recent_tools();
-        }
-        self.recent_tools()
+        session_runtime::recent_tools_for_session(self, _session)
     }
 
     pub(crate) fn record_file_access_for_session(
@@ -407,23 +253,14 @@ impl AppState {
         _session: &crate::session_context::SessionRequestContext,
         path: &str,
     ) {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            session_state.record_file_access(path);
-            return;
-        }
-        self.record_file_access(path);
+        session_runtime::record_file_access_for_session(self, _session, path);
     }
 
     pub(crate) fn recent_file_paths_for_session(
         &self,
         _session: &crate::session_context::SessionRequestContext,
     ) -> Vec<String> {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            return session_state.recent_file_paths();
-        }
-        self.recent_file_paths()
+        session_runtime::recent_file_paths_for_session(self, _session)
     }
 
     /// Returns (consecutive_count, is_rapid_burst).
@@ -435,20 +272,12 @@ impl AppState {
         name: &str,
         args_hash: u64,
     ) -> (usize, bool) {
-        #[cfg(feature = "http")]
-        if let Some(session_state) = self.http_session_state(_session) {
-            return session_state.doom_loop_count(name, args_hash);
-        }
-        self.doom_loop_count(name, args_hash)
+        session_runtime::doom_loop_count_for_session(self, _session, name, args_hash)
     }
 
     #[cfg(feature = "http")]
     pub(crate) fn bind_project_to_session(&self, session_id: &str, project_path: &str) {
-        if let Some(store) = &self.session_store
-            && let Some(session) = store.get(session_id)
-        {
-            session.set_project_path(project_path);
-        }
+        session_runtime::bind_project_to_session(self, session_id, project_path);
     }
 
     #[cfg(feature = "http")]
@@ -456,22 +285,7 @@ impl AppState {
         &'a self,
         session: &crate::session_context::SessionRequestContext,
     ) -> Result<Option<std::sync::MutexGuard<'a, ()>>, CodeLensError> {
-        let Some(bound_project) = session.project_path.as_deref() else {
-            return Ok(None);
-        };
-        let guard = self
-            .project_execution_lock
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        let current = self.current_project_scope();
-        if current != bound_project {
-            self.switch_project(bound_project).map_err(|error| {
-                CodeLensError::Validation(format!(
-                    "session project `{bound_project}` is not active and automatic rebind failed: {error}"
-                ))
-            })?;
-        }
-        Ok(Some(guard))
+        session_runtime::ensure_session_project(self, session)
     }
 
     #[cfg(not(feature = "http"))]
@@ -479,7 +293,7 @@ impl AppState {
         &'a self,
         _session: &crate::session_context::SessionRequestContext,
     ) -> Result<Option<std::sync::MutexGuard<'a, ()>>, CodeLensError> {
-        Ok(None)
+        session_runtime::ensure_session_project(self, _session)
     }
 
     pub(crate) fn preflight_ttl_seconds(&self) -> u64 {
@@ -633,42 +447,11 @@ impl AppState {
     }
 
     pub(crate) fn watcher_failure_health(&self) -> WatcherFailureHealth {
-        let symbol_index = self.symbol_index();
-        let db = symbol_index.db();
-        let summary = db
-            .index_failure_summary(WATCHER_RECENT_FAILURE_WINDOW_SECS)
-            .unwrap_or_default();
-        let scope = self.current_project_scope();
-        let pruned_missing_failures = self
-            .watcher_maintenance
-            .lock()
-            .unwrap_or_else(|p| p.into_inner())
-            .get(&scope)
-            .copied()
-            .unwrap_or(0);
-        WatcherFailureHealth {
-            recent_failures: summary.recent_failures,
-            total_failures: summary.total_failures,
-            stale_failures: summary.stale_failures,
-            persistent_failures: summary.persistent_failures,
-            pruned_missing_failures,
-            recent_window_seconds: WATCHER_RECENT_FAILURE_WINDOW_SECS,
-        }
+        watcher_health::watcher_failure_health(self)
     }
 
     pub(crate) fn prune_index_failures(&self) -> Result<WatcherFailureHealth, CodeLensError> {
-        let project = self.project();
-        let scope = self.current_project_scope();
-        let symbol_index = self.symbol_index();
-        let pruned_missing_failures = {
-            let db = symbol_index.db();
-            db.prune_missing_index_failures(project.as_path())?
-        };
-        self.watcher_maintenance
-            .lock()
-            .unwrap_or_else(|p| p.into_inner())
-            .insert(scope, pruned_missing_failures);
-        Ok(self.watcher_failure_health())
+        watcher_health::prune_index_failures(self)
     }
 
     /// Get the active graph cache.
